Bitwise Operators Considered Harmful
Changed most usage of bitwise operators (&, |, <<, >>) to use the bit()
or bits() methods of the Integer and Natural classes, severely dropping
emulation speed and causing collateral damage in the form of explicit
typecasts to cater to viewers who do not know the meanings of the bit
operators.

The performance profile will not compile in this version.

This is a change made by development versions of higan.
Note: Only code structural changes and bug fixes will be imported from
higan's betas.

+++ nSide/fc/arcade/pc10/pc10.cpp
@@ -54,7 +54,7 @@ auto PC10ArcadeBoard::read(uint16 addr) -> uint8 {
     if(!prom_test || prom_address < 0x40)
       data |= (((cartridge.board->keyrom.read(prom_address >> 3) >> (prom_address & 7)) & 1) << 3) ^ 0x08;
     else
-      data |= (prom_address & 2 ? 0x00 : cartridge.board->keyrom.read(8)) ^ 0x08;
+      data |= (prom_address & 2 ? (uint8)0x00 : cartridge.board->keyrom.read(8)) ^ 0x08;
     data |= ((prom_address & 0x20) >> 1) ^ 0x10;
     return data;
   };
